batch_training/requirements.txt
google-cloud-bigquery==3.3.1
db-dtypes==1.0.3
numpy==1.17.3
pandas==1.3.5
flake8==5.0.4
